hub/pyproject.toml
[tool.poetry]
name = "hub"
version = "0.1.0"
description = ""
authors = ["Zakaria Ouzarf <zakaria.ouzarf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
prisma = "^0.13.1"
pydantic = "^2.7.4"
paho-mqtt = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
